Build 2052beta

+++ skydrop/src/drivers/sensors/ms5611.h
@@ -56,7 +56,7 @@ public:
 
 	uint8_t address;
 
-	float temperature;
+	int32_t temperature;
 	float pressure;
 
 	uint32_t raw_temperature;
